+++ src/qmflows/common.py
@@ -47,8 +47,8 @@ class CGF(NamedTuple):
 class InfoMO(NamedTuple):
     """Energies and coefficients of the molecular orbitals."""
 
-    eigenVals: np.ndarray  # Orbitals eigenvalues
-    coeffs: np.ndarray  # Orbitals eigenvectors
+    eigenvalues: np.ndarray  # Orbitals eigenvalues
+    eigenvectors: np.ndarray  # Orbitals eigenvectors
 
 
 class MO_metadata(NamedTuple):